move FutureStream interface to api
