+++ src/main/java/com/aol/simple/react/stream/SimpleReact.java
@@ -21,6 +21,7 @@ import com.aol.simple.react.generators.Generator;
 import com.aol.simple.react.generators.ParallelGenerator;
 import com.aol.simple.react.generators.ReactIterator;
 import com.aol.simple.react.generators.SequentialIterator;
+import com.aol.simple.react.stream.api.FutureStream;
 import com.google.common.annotations.VisibleForTesting;
 import com.nurkiewicz.asyncretry.RetryExecutor;
