+++ src/main/java/com/aol/simple/react/stream/api/SyncToQueue.java
@@ -8,7 +8,6 @@ import java.util.stream.Collector;
 import com.aol.simple.react.async.Queue;
 import com.aol.simple.react.async.QueueFactories;
 import com.aol.simple.react.async.QueueFactory;
-import com.aol.simple.react.stream.FutureStream;
 
 public interface SyncToQueue <U> extends ToQueue<U>{
