+++ src/test/java/com/aol/simple/react/BaseSeqTest.java
@@ -28,7 +28,7 @@ import org.junit.Ignore;
 import org.junit.Test;
 
 import com.aol.simple.react.stream.EagerFutureStream;
-import com.aol.simple.react.stream.FutureStream;
+import com.aol.simple.react.stream.api.FutureStream;
 
 public abstract class BaseSeqTest {
 	abstract <U> FutureStream<U> of(U... array);
