+++ src/test/java/com/aol/simple/react/BaseSequentialSeqTest.java
@@ -27,7 +27,7 @@ import org.junit.Ignore;
 import org.junit.Test;
 
 import com.aol.simple.react.stream.EagerFutureStream;
-import com.aol.simple.react.stream.FutureStream;
+import com.aol.simple.react.stream.api.FutureStream;
 
 public abstract class BaseSequentialSeqTest {
 	abstract <U> FutureStream<U> of(U... array);
