+++ src/test/java/com/aol/simple/react/BlockingTest.java
@@ -11,8 +11,8 @@ import java.util.stream.Collectors;
 
 import org.junit.Test;
 
-import com.aol.simple.react.stream.FutureStream;
 import com.aol.simple.react.stream.SimpleReact;
+import com.aol.simple.react.stream.api.FutureStream;
 
 public class BlockingTest {
