+++ src/test/java/com/aol/simple/react/EagerSeqTest.java
@@ -10,7 +10,7 @@ import java.util.stream.Stream;
 import org.junit.Test;
 
 import com.aol.simple.react.stream.EagerFutureStream;
-import com.aol.simple.react.stream.FutureStream;
+import com.aol.simple.react.stream.api.FutureStream;
 
 public class EagerSeqTest extends BaseSeqTest {
