+++ src/test/java/com/aol/simple/react/MergeTest.java
@@ -8,8 +8,8 @@ import java.util.concurrent.ExecutionException;
 
 import org.junit.Test;
 
-import com.aol.simple.react.stream.FutureStream;
 import com.aol.simple.react.stream.SimpleReact;
+import com.aol.simple.react.stream.api.FutureStream;
 
 public class MergeTest {
