+++ src/test/java/com/aol/simple/react/completablefuture/CompletableFutureTest.java
@@ -15,8 +15,8 @@ import java.util.stream.Stream;
 
 import org.junit.Test;
 
-import com.aol.simple.react.stream.FutureStream;
 import com.aol.simple.react.stream.SimpleReact;
+import com.aol.simple.react.stream.api.FutureStream;
 
 public class CompletableFutureTest {
 	@Test
